Make interop stat tile values clickable on home
Top chain links to its project page, top token to its token page, and
top chain path to the interactive interop summary — matching the links
on /interop/summary. Chain hrefs are now populated in getHomeData via
the shared getInteropChainHref helper.
